ADD: APIs now require Authentication

+++ backend/src/core/views.py
@@ -46,24 +46,45 @@ def log_out_user(request):
 
 
 @api_view(["GET"])
-def get_character_overview(_, user_id):
-    character_overview_data = Character.objects.filter(user_id=user_id)
+def get_character_overview(request):
+    # check if user is authenticated and session is not expired
+    if not request.user.is_authenticated or not request.session.session_key:
+        return Response(status=403)
+
+    character_overview_data = Character.objects.filter(user_id=request.user.id)
     response_data = CharacterOverviewSerializer(character_overview_data, many=True).data
 
     return Response(response_data)
 
 
 @api_view(["GET"])
-def get_character(_, character_id):
-    character_data = Character.objects.get(pk=character_id)
-    response_data = CharacterSerializer(character_data).data
+def get_character(request, character_id):
+    # check if user is authenticated and session is not expired
+    if not request.user.is_authenticated or not request.session.session_key:
+        return Response(status=403)
 
-    return Response(response_data)
+    character = Character.objects.get(pk=character_id)
+
+    # check if the character belongs to the user
+    if character.user_id.id == request.user.id:
+        response_data = CharacterSerializer(character).data
+        return Response(response_data)
+
+    return Response(status=403)
 
 
 @api_view(["POST"])
 def save_character(request, character_id):
+    # check if user is authenticated and session is not expired
+    if not request.user.is_authenticated or not request.session.session_key:
+        return Response(status=403)
+
     character = Character.objects.filter(pk=character_id)
+
+    # check if the character belongs to the user
+    if character.user_id.id != request.user.id:
+        return Response(status=403)
+
     if character:
         character.update(**request.data)
     else:
@@ -74,11 +95,18 @@ def save_character(request, character_id):
 
 @api_view(["POST"])
 def create_character(request):
-    character_name = request.data["name"]
-    user_id = request.data["user_id"]
+    print(request.user)
+    print(request.user.id)
 
-    serializer = CharacterSerializer(data={"user_id": user_id, "name": character_name})
+    # check if user is authenticated and session is not expired
+    if not request.user.is_authenticated or not request.session.session_key:
+        return Response(status=403)
+
+    character_name = request.data["name"]
+
+    serializer = CharacterSerializer(data={"user_id": request.user.id, "name": character_name})
     if serializer.is_valid():
+        print(serializer.validated_data)
         serializer.save()
     else:
         return Response(status=404)
@@ -87,8 +115,17 @@ def create_character(request):
 
 
 @api_view(["POST"])
-def delete_character(_, character_id):
+def delete_character(request, character_id):
+    # check if user is authenticated and session is not expired
+    if not request.user.is_authenticated or not request.session.session_key:
+        return Response(status=403)
+
     character = Character.objects.filter(pk=character_id)
+
+    # check if the character belongs to the user
+    if character.user_id.id != request.user.id:
+        return Response(status=403)
+
     if character:
         character.delete()
     else:
